requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
boto3==1.42.77
botocore==1.42.77
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.8
cryptography==46.0.6
cssselect==1.4.0
distro==1.9.0
Django==5.2.11
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-filter==25.2
django-prometheus==2.4.1
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.25.2
flower==2.0.1
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.1
groq==1.1.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanize==4.15.0
idna==3.11
inflection==0.5.1
jieba3k==0.35.1
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
kombu==5.6.2
lxml==6.0.2
lxml_html_clean==0.4.4
newspaper3k==0.2.8
nltk==3.9.4
oauthlib==3.3.1
openai==2.28.0
packaging==26.0
pillow==12.1.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.2.0
regex==2026.2.28
requests==2.33.0
requests-file==3.0.1
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.56.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
tinysegmenter==0.3
tldextract==5.3.1
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
